restructure uiDeviceConfig classes
