+++ kaspresso/src/main/kotlin/com/kaspersky/kaspresso/kaspresso/KautomatorConfig.kt
@@ -1,7 +1,7 @@
 package com.kaspersky.kaspresso.kaspresso
 
 import com.kaspersky.kaspresso.device.Device
-import com.kaspersky.kaspresso.device.UiDeviceConfig
+import com.kaspersky.kaspresso.device.uideviceconfig.UiDeviceConfig
 import com.kaspersky.kaspresso.device.accessibility.Accessibility
 import com.kaspersky.kaspresso.device.activities.Activities
 import com.kaspersky.kaspresso.device.apps.Apps
